format
1

+++ editor/builtin/cloud-function/panel/utils/tcb_utils.js
@@ -1,19 +1,19 @@
-const n = require("@cloudbase/manager-node"),
-  t = require("fs"),
-  e = require("./utils"),
-  a = require("compressing"),
-  i = require("request");
-const r =
-    '\n{\n  "name": "#{func_name}",\n  "version": "1.0.0",\n  "main": "index.js",\n  "dependencies": {\n    "@cloudbase/node-sdk": "^2.5.0"\n  }\n}\n',
-  o =
-    "\n'use strict';\nconst tcb = require('@cloudbase/node-sdk');\nconst app = tcb.init();\nconst auth = app.auth();\nconst db = app.database();\nexports.main = async (event, context) => {\n  console.log(\"Hello World\");\n  console.log(event);\n  console.log(event[\"non-exist\"]);\n  console.log(context);\n  return event;\n};\n",
-  c =
-    "\nResources:\n  #{env_id}:\n    Type: TencentCloud::Serverless::Namespace\n    #{func_name}:\n      Properties:\n        CodeUri: .\n        Description: ''\n        Environment:\n          Variables: {}\n        Events: {}\n        Handler: index.main\n        MemorySize: 256\n        Role: TCB_QcsRole\n        Runtime: Nodejs10.15\n        Timeout: 20\n        VpcConfig:\n          SubnetId: ''\n          VpcId: ''\n      Type: TencentCloud::Serverless::Function\n";
+const n = require("@cloudbase/manager-node");
+const t = require("fs");
+const e = require("./utils");
+const a = require("compressing");
+const i = require("request");
+const r = '\n{\n  "name": "#{func_name}",\n  "version": "1.0.0",\n  "main": "index.js",\n  "dependencies": {\n    "@cloudbase/node-sdk": "^2.5.0"\n  }\n}\n';
+const o = "\n'use strict';\nconst tcb = require('@cloudbase/node-sdk');\nconst app = tcb.init();\nconst auth = app.auth();\nconst db = app.database();\nexports.main = async (event, context) => {\n  console.log(\"Hello World\");\n  console.log(event);\n  console.log(event[\"non-exist\"]);\n  console.log(context);\n  return event;\n};\n";
+const c = "\nResources:\n  #{env_id}:\n    Type: TencentCloud::Serverless::Namespace\n    #{func_name}:\n      Properties:\n        CodeUri: .\n        Description: ''\n        Environment:\n          Variables: {}\n        Events: {}\n        Handler: index.main\n        MemorySize: 256\n        Role: TCB_QcsRole\n        Runtime: Nodejs10.15\n        Timeout: 20\n        VpcConfig:\n          SubnetId: ''\n          VpcId: ''\n      Type: TencentCloud::Serverless::Function\n";
+
 module.exports = {
-  inited: !1,
+  inited: false,
   app: {},
   onError(n) {
-    this.onTCBError && this.onTCBError(n);
+    if (this.onTCBError) {
+      this.onTCBError(n);
+    }
   },
   async init(t) {
     this.app = null;
@@ -23,12 +23,17 @@ module.exports = {
       return;
     }
     var a = await this.app.env.listEnvs().catch((n) => this.onError(n));
-    if ((this.app.getEnvironmentManager().remove(""), a)) {
+    this.app.getEnvironmentManager().remove("");
+    if ((a)) {
       for (var i of a.EnvList) this.app.getEnvironmentManager().add(i.EnvId);
       var r = e.getCurrentEnvId();
-      "undefinedenv" === r && (r = a.EnvList[0] && a.EnvList[0].EnvId),
-        this.app.getEnvironmentManager().switchEnv(r),
-        (this.inited = !0);
+
+      if ("undefinedenv" === r) {
+        r = a.EnvList[0] && a.EnvList[0].EnvId;
+      }
+
+      this.app.getEnvironmentManager().switchEnv(r);
+      this.inited = true;
     }
   },
   switchEnv(n) {
@@ -91,15 +96,19 @@ module.exports = {
     ) {
       (function (n, a, i) {
         var s = `${n}/${a}`;
-        t.existsSync(s) ||
-          (e.mkdirs(s),
-          t.writeFileSync(`${s}/package.json`, r.replace(/#{func_name}/g, a)),
-          t.writeFileSync(`${s}/index.js`, o.replace(/#{env_id}/g, i)),
+
+        if (!t.existsSync(s)) {
+          e.mkdirs(s);
+          t.writeFileSync(`${s}/package.json`, r.replace(/#{func_name}/g, a));
+          t.writeFileSync(`${s}/index.js`, o.replace(/#{env_id}/g, i));
+
           t.writeFileSync(
             `${s}/template.ymal`,
             c.replace(/#{env_id}/g, i).replace(/#{func_name}/g, a)
-          ),
-          console.log("创建成功"));
+          );
+
+          console.log("创建成功");
+        }
       })(a, n.name, u.EnvInfo.EnvId);
       var p = await this.app.functions
         .createFunction({
@@ -109,52 +118,69 @@ module.exports = {
           base64Code: s,
         })
         .catch((n) => this.onError(n));
-      return await this.downloadFunction(n.name, a), { res: p, isCreate: !0 };
+      await this.downloadFunction(n.name, a);
+      return { res: p, isCreate: true };
     }
-    return { res: await this.downloadFunction(n.name, a), isCreate: !1 };
+    return { res: await this.downloadFunction(n.name, a), isCreate: false };
   },
   async downloadFunction(n, e) {
     var r = await this.app.functions
         .getFunctionDownloadUrl(n)
-        .catch((n) => this.onError(n)),
-      o = `${e}/${n}.zip`;
+        .catch((n) => this.onError(n));
+
+    var o = `${e}/${n}.zip`;
     return new Promise((c, s) => {
       (function (n, e, a) {
-        var r = 0,
-          o = 0,
-          c = 0,
-          s = i({
-            method: "GET",
-            uri: n,
-            agentOptions: { ciphers: "ALL", secureProtocol: "TLSv1_method" },
-          }),
-          u = t.createWriteStream(e);
-        s.pipe(u),
-          s.on("response", (n) => {
-            if (200 !== n.statusCode)
-              return a(new Error(n.statusMessage), null), void 0;
-            r = parseInt(n.headers["content-length"], 10);
-          }),
-          s.on("data", (n) => {
-            var t = (((o += n.length) / r) * 100) | 0;
-            c !== t &&
-              ((c = t), a(null, { status: "downloading", progress: t }));
-          }),
-          s.on("complete", () => {
-            a(null, { status: "complete", progress: c });
-          }),
-          s.on("error", (n) => {
-            a(n, null);
-          });
+        var r = 0;
+        var o = 0;
+        var c = 0;
+
+        var s = i({
+          method: "GET",
+          uri: n,
+          agentOptions: { ciphers: "ALL", secureProtocol: "TLSv1_method" },
+        });
+
+        var u = t.createWriteStream(e);
+        s.pipe(u);
+
+        s.on("response", (n) => {
+          if (200 !== n.statusCode) {
+            a(new Error(n.statusMessage), null);
+            return;
+          }
+          r = parseInt(n.headers["content-length"], 10);
+        });
+
+        s.on("data", (n) => {
+          var t = (((o += n.length) / r) * 100) | 0;
+
+          if (c !== t) {
+            c = t;
+            a(null, { status: "downloading", progress: t });
+          }
+        });
+
+        s.on("complete", () => {
+          a(null, { status: "complete", progress: c });
+        });
+
+        s.on("error", (n) => {
+          a(n, null);
+        });
       })(r.Url, o, (i, r) => {
-        i ||
-          "complete" !== r.status ||
+        if (!(i || "complete" !== r.status)) {
           a.zip
             .uncompress(o, `${e}/${n}/`)
             .then(() => {
-              t.existsSync(o) && t.unlinkSync(o), c();
-            })
+            if (t.existsSync(o)) {
+              t.unlinkSync(o);
+            }
+
+            c();
+          })
             .catch((n) => {});
+        }
       });
     });
   },
@@ -169,13 +195,13 @@ module.exports = {
               timeout: 20,
               envVariables: {},
               runtime: "Nodejs8.9",
-              installDependency: !0,
-              isWaitInstall: !0,
+              installDependency: true,
+              isWaitInstall: true,
               triggers: [],
               ignore: [],
             },
             functionRootPath: `${t}/`,
-            force: !0,
+            force: true,
             base64Code: e,
           })
           .catch((n) => this.onError(n))
@@ -208,7 +234,7 @@ module.exports = {
       .getFunctionLogs(n)
       .catch((n) => this.onError(n));
   },
-  async copyFunction(n, t, e = "", a = !1) {
+  async copyFunction(n, t, e = "", a = false) {
     return await this.app.functions
       .copyFunction(n, t, e, a)
       .catch((n) => this.onError(n));
@@ -224,58 +250,67 @@ module.exports = {
       .catch((n) => this.onError(n));
   },
   async createSafetySource(n, t = "") {
-    if ("undefinedenv" !== t)
-      return (
-        (null !== t && "" !== t) || (t = this.app.env.envId),
-        await this.app
-          .commonService()
-          .call({
-            Action: "CreateSafetySource",
-            Param: { EnvId: t, AppName: n },
-          })
-          .catch((n) => this.onError(n)),
-        (await this.describeSafetySource()).Data.filter(
-          (t) => t.AppName === n
-        )[0]
-      );
+    if ("undefinedenv" !== t) {
+      if (!(null !== t && "" !== t)) {
+        t = this.app.env.envId;
+      }
+
+      await this.app
+        .commonService()
+        .call({
+          Action: "CreateSafetySource",
+          Param: { EnvId: t, AppName: n },
+        })
+        .catch((n) => this.onError(n));
+
+      return (await this.describeSafetySource()).Data.filter(
+        (t) => t.AppName === n
+      )[0];
+    }
   },
   async describeSafetySource(n) {
-    if ("undefinedenv" !== n)
-      return (
-        (null !== n && "" !== n) || (n = this.app.env.envId),
-        await this.app
-          .commonService()
-          .call({
-            Action: "DescribeSafetySource",
-            Param: { EnvId: n, Offset: 0, Limit: 100 },
-          })
-          .catch((n) => this.onError(n))
-      );
+    if ("undefinedenv" !== n) {
+      if (!(null !== n && "" !== n)) {
+        n = this.app.env.envId;
+      }
+
+      return await this.app
+        .commonService()
+        .call({
+          Action: "DescribeSafetySource",
+          Param: { EnvId: n, Offset: 0, Limit: 100 },
+        })
+        .catch((n) => this.onError(n));
+    }
   },
   async describeSafetySourceSecretKey(n, t = "") {
-    if ("undefinedenv" !== t)
-      return (
-        (null !== t && "" !== t) || (t = this.app.env.envId),
-        await this.app
-          .commonService()
-          .call({
-            Action: "DescribeSafetySourceSecretKey",
-            Param: { EnvId: t, ItemId: n },
-          })
-          .catch((n) => this.onError(n))
-      );
+    if ("undefinedenv" !== t) {
+      if (!(null !== t && "" !== t)) {
+        t = this.app.env.envId;
+      }
+
+      return await this.app
+        .commonService()
+        .call({
+          Action: "DescribeSafetySourceSecretKey",
+          Param: { EnvId: t, ItemId: n },
+        })
+        .catch((n) => this.onError(n));
+    }
   },
   async deleteSafetySource(n, t = "") {
-    if ("undefinedenv" !== t)
-      return (
-        (null !== t && "" !== t) || (t = this.app.env.envId),
-        await this.app
-          .commonService()
-          .call({
-            Action: "DeleteSafetySource",
-            Param: { EnvId: t, ItemId: n },
-          })
-          .catch((n) => this.onError(n))
-      );
+    if ("undefinedenv" !== t) {
+      if (!(null !== t && "" !== t)) {
+        t = this.app.env.envId;
+      }
+
+      return await this.app
+        .commonService()
+        .call({
+          Action: "DeleteSafetySource",
+          Param: { EnvId: t, ItemId: n },
+        })
+        .catch((n) => this.onError(n));
+    }
   },
 };
