+++ editor/builtin/xiaomi-runtime/lib/killPort.js
@@ -1,13 +1,15 @@
 "use strict";
-var t = require("child_process"),
-  n = t.spawn,
-  o = (t.exec, t.execSync),
-  r = t.spawnSync;
+var t = require("child_process");
+var n = t.spawn;
+var o = (t.exec, t.execSync);
+var r = t.spawnSync;
 function i(t, n) {
   try {
     process.kill(parseInt(t, 10), n);
   } catch (t) {
-    if ("ESRCH" !== t.code) throw t;
+    if ("ESRCH" !== t.code) {
+      throw t;
+    }
   }
 }
 function e(t, n, o, r, i) {
@@ -18,7 +20,8 @@ function e(t, n, o, r, i) {
     a += t;
   });
   var f = function (c) {
-    if ((delete o[t], 0 != c)) {
+    delete o[t];
+    if ((0 != c)) {
       if (Object.keys(o).length == 0) {
         i();
       }
@@ -34,9 +37,10 @@ function e(t, n, o, r, i) {
   };
   c.on("close", f);
 }
+
 module.exports = function (t, c, a, f) {
-  var s = {},
-    u = {};
+  var s = {};
+  var u = {};
   switch (
     ((s[t] = []),
     (u[t] = 1),
@@ -49,22 +53,28 @@ module.exports = function (t, c, a, f) {
     case "darwin":
       (function (t) {
         var n;
-        if (
-          ((function (t) {
-            for (var n = "", o = 0; o < t.length; o++)
+
+        (function (t) {
+            for (var n = "", o = 0; o < t.length; o++) {
               n += String.fromCharCode(t[o]);
+            }
             return n;
           })(r("lsof", ["-i", `:${t}`]).stdout)
             .split(/[\n|\r]/)
             .forEach((t) => {
               if (-1 !== t.indexOf("LISTEN") && !n) {
                 let o = t.split(/\s+/);
-                /\d+/.test(o[1]) && (n = o[1]);
+
+                if (/\d+/.test(o[1])) {
+                  n = o[1];
+                }
               }
-            }),
-          !n)
-        )
-          return Editor.log(`port:${t} close!`), void 0;
+            });
+
+        if (!n) {
+          Editor.log(`port:${t} close!`);
+          return;
+        }
         o(`kill -9 ${n}`);
       })(c);
       break;
@@ -77,32 +87,45 @@ module.exports = function (t, c, a, f) {
           a += t;
         });
         c.on("close", function (c) {
-          if ((delete o[t], 0 != c))
-            return 0 == Object.keys(o).length && i(), void 0;
+          delete o[t];
+          if (0 != c) {
+            if (0 == Object.keys(o).length) {
+              i();
+            }
+
+            return;
+          }
           a.match(/\d+/g).forEach(function (c) {
-            (c = parseInt(c, 10)),
-              n[t].push(c),
-              (n[c] = []),
-              (o[c] = 1),
-              (function (t, n, o, r, i) {
-                var c = r(t);
-                var a = "";
-                c.stdout.on("data", function (t) {
-                  var t = t.toString("ascii");
-                  a += t;
+            c = parseInt(c, 10);
+            n[t].push(c);
+            n[c] = [];
+            o[c] = 1;
+
+            (function (t, n, o, r, i) {
+              var c = r(t);
+              var a = "";
+              c.stdout.on("data", function (t) {
+                var t = t.toString("ascii");
+                a += t;
+              });
+              c.on("close", function (c) {
+                delete o[t];
+                if (0 != c) {
+                  if (0 == Object.keys(o).length) {
+                    i();
+                  }
+
+                  return;
+                }
+                a.match(/\d+/g).forEach(function (c) {
+                  c = parseInt(c, 10);
+                  n[t].push(c);
+                  n[c] = [];
+                  o[c] = 1;
+                  e(c, n, o, r, i);
                 });
-                c.on("close", function (c) {
-                  if ((delete o[t], 0 != c))
-                    return 0 == Object.keys(o).length && i(), void 0;
-                  a.match(/\d+/g).forEach(function (c) {
-                    (c = parseInt(c, 10)),
-                      n[t].push(c),
-                      (n[c] = []),
-                      (o[c] = 1),
-                      e(c, n, o, r, i);
-                  });
-                });
-              })(c, n, o, r, i);
+              });
+            })(c, n, o, r, i);
           });
         });
       })(
@@ -118,15 +141,26 @@ module.exports = function (t, c, a, f) {
             try {
               Object.keys(t).forEach(function (o) {
                 t[o].forEach(function (t) {
-                  r[t] || (i(t, n), (r[t] = 1));
-                }),
-                  r[o] || (i(o, n), (r[o] = 1));
+                  if (!r[t]) {
+                    i(t, n);
+                    r[t] = 1;
+                  }
+                });
+
+                if (!r[o]) {
+                  i(o, n);
+                  r[o] = 1;
+                }
               });
             } catch (t) {
-              if (o) return o(t);
+              if (o) {
+                return o(t);
+              }
               throw t;
             }
-            if (o) return o();
+            if (o) {
+              return o();
+            }
           })(s, a, f);
         }
       );
